sys-src/GetraenkeIO-backend/requirements.txt
fastapi[standard]
sqlmodel
pydantic-settings
python-dotenv
passlib
bcrypt
pyjwt
httpx
pytest